% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_Prediction.R
\name{compare_Prediction}
\alias{compare_Prediction}
\title{Compare Predictive Models Over Future Periods}
\usage{
compare_Prediction(
  start_d,
  end_d,
  case,
  GT,
  pred_period,
  pred_length = 1,
  lags = NULL,
  pred_unit = "Month",
  offset_GT = 0.001
)
}
\arguments{
\item{start_d}{The start date for the overall period considered for model training, formatted as "YYYY-MM-DD".}

\item{end_d}{The end date for the overall period considered for model training, formatted as "YYYY-MM-DD".}

\item{case}{A data frame containing case data, expected to include at least two columns: "date" and "N_cases".}

\item{GT}{A data frame containing Google Trends data relevant to the cases under study.}

\item{pred_period}{The total number of separate future periods for which predictions will be made.}

\item{pred_length}{The forecast horizon for each prediction, specified in months or weeks.
Valid options are 1 or 2, indicating the length of time to forecast ahead.}

\item{lags}{An optional numeric vector specifying the lag periods to be applied when generating
lagged data for model inputs. The default settings may vary based on the selected model.}

\item{pred_unit}{The unit of measurement for the prediction period, with options "Month" or "Week".
The default is "Month".}

\item{offset_GT}{A small positive numeric added to Google Trends data before logarithmic transformation
to manage instances of zero. The default value is 0.001.}
}
\value{
A data frame containing columns for actual case counts ('Real_value') and the predicted case
        counts from each of the specified models, including a column for naive model predictions.
        Each row corresponds to a prediction for a future period as defined by `pred_period` and
        `pred_length`.
}
\description{
This function performs a comparison of forecasts generated by multiple predictive models against
actual case data over designated future periods. It uses case data along with Google Trends data
as inputs and evaluates the performance of models such as ARGO, SAR, GT, and SAR_GT. Additionally,
a naive prediction based on the most recent period's data is included for comparison.
}
\examples{
compare_Prediction("2020-01-01", "2020-06-30", case_data, GT_data, 2, pred_length = 1)
}
