% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DengueARGO-package.R
\docType{package}
\name{DengueARGO}
\alias{DengueARGO-package}
\alias{DengueARGO}
\title{DengueARGO Package}
\description{
DengueARGO package which provides functions
to predict disease spread using models such as ARGO, SAR, and GT. It
allows users to generate lagged data sets, construct model matrices, fit
models, and evaluate predictions.
}
\author{
Yang Xiao \email{yang.xiao@kaust.edu.sa}
}
