% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Box.R
\name{get_Boxplot}
\alias{get_Boxplot}
\title{Generate Boxplot for Model Prediction Errors}
\usage{
get_Boxplot(
  data,
  type = "Diff",
  order_by = "Model",
  plot_title = "Difference between Predictions and Real Value",
  x_lab = "Prediction Method",
  y_lab = "Difference"
)
}
\arguments{
\item{data}{A data frame with the first column containing the actual values and subsequent 
columns containing predictions from different models. Column names should represent model names.}

\item{type}{Specifies whether to calculate absolute differences ("Diff") or percentage differences ("Perc") 
between predicted and actual values. Default is "Diff".}

\item{order_by}{Determines the ordering of models in the boxplot, either by "Model" names or by the 
"Median" value of their prediction errors. Default is "Model".}

\item{plot_title}{Title for the boxplot. Default is "Difference between Predictions and Real Value".}

\item{x_lab}{Label for the x-axis, typically representing the prediction method. Default is "Prediction Method".}

\item{y_lab}{Label for the y-axis, representing the type of difference (absolute or percentage). 
Default is "Difference".}
}
\value{
A `ggplot` object representing the boxplot of prediction errors for each model. The boxplot 
        can be customized further by the user using `ggplot2` functions.
}
\description{
Creates a boxplot visualizing the distribution of prediction errors (differences or percentages) 
for various predictive models compared to actual values. This function allows for a direct 
visual comparison of model performance in terms of prediction accuracy. Models can be ordered 
by their names or the median of their prediction errors in the plot.
}
\examples{
# Assuming `predictions_data` is your data frame with actual values in the first column and
# predictions from various models in subsequent columns:
plot <- get_Boxplot(predictions_data, type = "Diff", order_by = "Median")
plot
}
