% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Measurement.R
\name{get_Measurement}
\alias{get_Measurement}
\title{Calculate Prediction Performance Measurements}
\usage{
get_Measurement(y_fitted, y, IF_log = F)
}
\arguments{
\item{y_fitted}{The fitted (predicted) values.}

\item{y}{The actual values.}

\item{IF_log}{A logical indicating whether the values were transformed using logarithm. Default is FALSE.}
}
\value{
A data frame containing the performance measurements: RMSE, MAE, RMSPE, MAPE, and CORR.
}
\description{
This function calculates various performance measurements for the fitted (predicted) vs. actual data.
}
\examples{
get_Measurement(fitted_values, actual_values)
}
