% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Prediction.R
\name{get_Prediction}
\alias{get_Prediction}
\title{Fit Model and Make Prediction}
\usage{
get_Prediction(
  start_train,
  end_train,
  start_pred,
  end_pred,
  model = "ARGO",
  case,
  GT,
  lags = NULL,
  pred_unit = "M",
  offset_GT = 0.001
)
}
\arguments{
\item{start_train}{The start date for the training period.}

\item{end_train}{The end date for the training period.}

\item{start_pred}{The start date for the prediction period.}

\item{end_pred}{The end date for the prediction period.}

\item{model}{The model to fit ("ARGO", "SAR", "GT", or "SAR_GT").}

\item{case}{A data frame containing case data.}

\item{GT}{A data frame containing Google Trends data.}
}
\value{
A data frame with actual and predicted cases for the prediction period.
}
\description{
This function fits the specified prediction model to the training data and makes predictions for the given period.
}
\examples{
get_Prediction("ARGO", "2020-01-01", "2020-06-30", "2020-07-01", "2020-07-31", case_data, GT_data)
}
