% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_X.R
\name{get_X}
\alias{get_X}
\title{Get the X Matrix for Model Fitting}
\usage{
get_X(start, end, case, GT, model)
}
\arguments{
\item{start}{The start date for generating the X matrix, in "YYYY-MM-DD" format.}

\item{end}{The end date for generating the X matrix, in "YYYY-MM-DD" format.}

\item{case}{A data frame of case data, with Dengue case variable named "N_cases" and date variable named "date".}

\item{GT}{A data frame of Google Trends data.}

\item{model}{A string specifying the model type ("ARGO", "SAR", "GT", or "SAR_GT").}
}
\value{
A matrix X used for fitting the specified model.
}
\description{
This function generates the X matrix required for different prediction models by processing case and Google Trends (GT) data.
}
\examples{
get_X("2020-01-01", "2020-12-31", case_data, GT_data, "ARGO")
}
