% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_Gen.R
\name{lag_Gen}
\alias{lag_Gen}
\title{Produce Lagged Data}
\usage{
lag_Gen(data, start, end, lag_m)
}
\arguments{
\item{data}{A data frame containing the data to be lagged.}

\item{start}{The start date from which to begin lagging, in "YYYY-MM-DD" format.}

\item{end}{The end date for the lagging period, in "YYYY-MM-DD" format.}

\item{lag_m}{The number of months to lag by.}
}
\value{
A vector containing the lagged N_cases for the specified period.
}
\description{
This function produces lagged data for the specified period and lag months.
}
\examples{
lag_Gen(data, "2020-01-01", "2020-12-31", 3)
}
