% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_Measurement.R
\name{pred_Measurement}
\alias{pred_Measurement}
\title{Aggregate Prediction and Measurement for Multiple Models and Periods}
\usage{
pred_Measurement(start_d, end_d, pred_period, pred_length = 1)
}
\arguments{
\item{start_d}{The start date for the overall prediction period.}

\item{end_d}{The end date for the overall prediction period.}

\item{pred_period}{The total number of prediction periods.}

\item{pred_length}{The length of each prediction period in months (either 1 or 2).}
}
\value{
A data frame of performance measurements for each model and prediction period.
}
\description{
This function aggregates predictions and calculates performance measurements across multiple models and prediction periods.
}
\examples{
pred_Measurement("2020-01-01", "2020-12-31", 12, 1)
}
